requirements.txt
pandas
google-genai
python-dotenv
Flask
gunicorn  